// src/client/models/CheckMaterialMatchEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * - MATCHED_MATERIAL_PACK: 可在erp查到物料
 * - TESTING_MATERIAL_PACK: 廠商提供測試物料
 * - UNMATCHED_MATERIAL_PACK: 查不到或放錯槽位
 * - UNCHECKED_MATERIAL_PACK: 未經確認的物料
 */
export enum CheckMaterialMatchEnum {
    MATCHED_MATERIAL_PACK = 'MATCHED_MATERIAL_PACK',
    TESTING_MATERIAL_PACK = 'TESTING_MATERIAL_PACK',
    UNMATCHED_MATERIAL_PACK = 'UNMATCHED_MATERIAL_PACK',
    UNCHECKED_MATERIAL_PACK = 'UNCHECKED_MATERIAL_PACK',
}